Move single asset invest logic to InvestForm component
