--- src/components/forms/pool_actions/InvestForm/InvestSingleAssetForm.vue
@@ -1,270 +1,0 @@
-<script setup lang="ts">
-import { computed, onBeforeMount, ref, toRef, watch } from 'vue';
-import { useI18n } from 'vue-i18n';
-
-import WrapStEthLink from '@/components/contextual/pages/pool/invest/WrapStEthLink.vue';
-import StakePreviewModal from '@/components/contextual/stake/StakePreviewModal.vue';
-import TokenInput from '@/components/inputs/TokenInput/TokenInput.vue';
-import usePoolTransfers from '@/composables/contextual/pool-transfers/usePoolTransfers';
-import { isStablePhantom, usePool } from '@/composables/usePool';
-import useTokens from '@/composables/useTokens';
-import useVeBal from '@/composables/useVeBAL';
-import { LOW_LIQUIDITY_THRESHOLD } from '@/constants/poolLiquidity';
-import { bnum } from '@/lib/utils';
-import { isRequired } from '@/lib/utils/validations';
-import StakingProvider from '@/providers/local/staking/staking.provider';
-import { Pool } from '@/services/pool/types';
-import useWeb3 from '@/services/web3/useWeb3';
-
-import InvestFormTotals from './components/InvestFormTotals.vue';
-import InvestPreviewModal from './components/InvestPreviewModal/InvestPreviewModal.vue';
-import useInvestMath from './composables/useInvestMath';
-import useInvestState from './composables/useInvestState';
-
-/**
- * TYPES
- */
-enum NativeAsset {
-  wrapped = 'wrapped',
-  unwrapped = 'unwrapped'
-}
-
-type Props = {
-  pool: Pool;
-};
-
-/**
- * PROPS & EMITS
- */
-const props = defineProps<Props>();
-
-/**
- * STATE
- */
-const showInvestPreview = ref(false);
-const showStakeModal = ref(false);
-const pool = toRef(props, 'pool');
-
-/**
- * COMPOSABLES
- */
-
-const { balanceFor, nativeAsset, wrappedNativeAsset } = useTokens();
-const { useNativeAsset } = usePoolTransfers();
-const {
-  tokenAddresses,
-  validInputs,
-  highPriceImpactAccepted,
-  resetAmounts,
-  amounts
-} = useInvestState();
-const { t } = useI18n();
-
-const investMath = useInvestMath(pool, tokenAddresses, amounts, useNativeAsset);
-
-const { hasAmounts, highPriceImpact, swapRouteLoading, swapRoute } = investMath;
-
-const {
-  isWalletReady,
-  startConnectWithInjectedProvider,
-  isMismatchedNetwork
-} = useWeb3();
-
-const { managedPoolWithTradingHalted, isWethPool } = usePool(pool);
-const { veBalTokenInfo } = useVeBal();
-
-/**
- * COMPUTED
- */
-const error = computed(() => {
-  if (hasAmounts.value && swapRoute.value?.returnAmountFromSwaps.eq(0)) {
-    return {
-      header: t('insufficientLiquidity'),
-      body: t('insufficientLiquidityDetailed')
-    };
-  }
-  return null;
-});
-
-const hasValidInputs = computed(
-  (): boolean =>
-    validInputs.value.every(validInput => validInput === true) &&
-    hasAcceptedHighPriceImpact.value
-);
-
-const hasAcceptedHighPriceImpact = computed((): boolean =>
-  highPriceImpact.value ? highPriceImpactAccepted.value : true
-);
-
-const forceProportionalInputs = computed(
-  (): boolean => managedPoolWithTradingHalted.value
-);
-
-const poolHasLowLiquidity = computed((): boolean =>
-  bnum(props.pool.totalLiquidity).lt(LOW_LIQUIDITY_THRESHOLD)
-);
-
-const investmentTokens = computed((): string[] => {
-  if (isStablePhantom(props.pool.poolType)) {
-    return props.pool.mainTokens || [];
-  }
-  return props.pool.tokensList;
-});
-
-// If ETH has a higher balance than WETH then use it for the input.
-function setNativeAssetByBalance(): void {
-  const nativeAssetBalance = balanceFor(nativeAsset.address);
-  const wrappedNativeAssetBalance = balanceFor(
-    wrappedNativeAsset.value.address
-  );
-
-  if (bnum(nativeAssetBalance).gt(wrappedNativeAssetBalance)) {
-    setNativeAsset(NativeAsset.unwrapped);
-    useNativeAsset.value = true;
-  }
-}
-
-function setNativeAsset(to: NativeAsset): void {
-  const fromAddress =
-    to === NativeAsset.wrapped
-      ? nativeAsset.address
-      : wrappedNativeAsset.value.address;
-  const toAddress =
-    to === NativeAsset.wrapped
-      ? wrappedNativeAsset.value.address
-      : nativeAsset.address;
-
-  const indexOfAsset = tokenAddresses.value.indexOf(fromAddress);
-
-  if (indexOfAsset >= 0) {
-    tokenAddresses.value[indexOfAsset] = toAddress;
-  }
-}
-
-/**
- * CALLBACKS
- */
-onBeforeMount(() => {
-  resetAmounts();
-  tokenAddresses.value = [...investmentTokens.value];
-  if (isWethPool.value) setNativeAssetByBalance();
-});
-
-/**
- * WATCHERS
- */
-watch(useNativeAsset, shouldUseNativeAsset => {
-  if (shouldUseNativeAsset) {
-    setNativeAsset(NativeAsset.unwrapped);
-  } else {
-    setNativeAsset(NativeAsset.wrapped);
-  }
-});
-</script>
-
-<template>
-  <div>
-    <BalAlert
-      v-if="forceProportionalInputs"
-      type="warning"
-      :title="$t('investment.warning.managedPoolTradingHalted.title')"
-      :description="
-        $t('investment.warning.managedPoolTradingHalted.description')
-      "
-      class="mb-4"
-    />
-
-    <BalAlert
-      v-if="poolHasLowLiquidity"
-      type="warning"
-      :title="$t('investment.warning.lowLiquidity.title')"
-      :description="$t('investment.warning.lowLiquidity.description')"
-      class="mb-4"
-    />
-
-    <TokenInput
-      :amount="amounts[0]"
-      :address="tokenAddresses[0]"
-      name="tokenIn"
-      class="mb-4"
-      @update:amount="amount => (amounts[0] = amount)"
-      @update:address="address => (tokenAddresses[0] = address)"
-      :excludedTokens="[veBalTokenInfo?.address, props.pool.address]"
-    />
-
-    <InvestFormTotals
-      :math="investMath"
-      @maximize="() => {}"
-      @optimize="() => {}"
-    />
-
-    <div
-      v-if="highPriceImpact"
-      class="border dark:border-gray-700 rounded-lg p-2 pb-2 mt-4"
-    >
-      <BalCheckbox
-        v-model="highPriceImpactAccepted"
-        :rules="[isRequired($t('priceImpactCheckbox'))]"
-        name="highPriceImpactAccepted"
-        size="sm"
-        :label="$t('priceImpactAccept', [$t('depositing')])"
-      />
-    </div>
-
-    <WrapStEthLink :pool="pool" class="mt-4" />
-
-    <div class="mt-4">
-      <BalAlert
-        v-if="error"
-        class="p-3 mb-4"
-        type="error"
-        size="sm"
-        :title="error.header"
-        :description="error.body"
-        block
-      />
-      <BalBtn
-        v-if="!isWalletReady"
-        :label="$t('connectWallet')"
-        color="gradient"
-        block
-        @click="startConnectWithInjectedProvider"
-      />
-      <BalBtn
-        v-else
-        :label="$t('preview')"
-        color="gradient"
-        :loading="swapRouteLoading"
-        :loading-label="$t('loadingBestPrice')"
-        :disabled="
-          !!error ||
-            !hasAmounts ||
-            !hasValidInputs ||
-            isMismatchedNetwork ||
-            swapRouteLoading
-        "
-        block
-        @click="showInvestPreview = true"
-      />
-    </div>
-
-    <StakingProvider :poolAddress="pool.address">
-      <teleport to="#modal">
-        <InvestPreviewModal
-          v-if="showInvestPreview"
-          :pool="pool"
-          :math="investMath"
-          :tokenAddresses="tokenAddresses"
-          @close="showInvestPreview = false"
-          @showStakeModal="showStakeModal = true"
-        />
-        <StakePreviewModal
-          :pool="pool"
-          :isVisible="showStakeModal"
-          @close="showStakeModal = false"
-          action="stake"
-        />
-      </teleport>
-    </StakingProvider>
-  </div>
-</template>
