+++ src/pages/pool/invest.vue
@@ -4,7 +4,6 @@ import { onBeforeMount } from 'vue';
 import useInvestState from '@/components/forms/pool_actions/InvestForm/composables/useInvestState';
 // Components
 import InvestForm from '@/components/forms/pool_actions/InvestForm/InvestForm.vue';
-import InvestSingleAssetForm from '@/components/forms/pool_actions/InvestForm/InvestSingleAssetForm.vue';
 import TradeSettingsPopover, {
   TradeSettingsContext,
 } from '@/components/popovers/TradeSettingsPopover.vue';
@@ -57,8 +56,7 @@ onBeforeMount(async () => {
           </div>
         </div>
       </template>
-      <InvestSingleAssetForm v-if="isStablePhantomPool" :pool="pool" />
-      <InvestForm v-else :pool="pool" />
+      <InvestForm :pool="pool" />
     </BalCard>
   </div>
 </template>
